refactor(host): split Beads lifecycle internals
Separate host-infra-system Beads path ownership from shared Dolt server coordination and add a dedicated lifecycle error model in host-infra-beads. This leaves the extracted lifecycle boundaries easier to reason about, test, and extend without mixing path resolution, process orchestration, and failure reporting.
