+++ apps/desktop/src-tauri/crates/host-infra-beads/Cargo.toml
@@ -8,5 +8,6 @@ anyhow = "1"
 chrono = { version = "0.4", features = ["serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+thiserror = "2"
 host-domain = { path = "../host-domain" }
 host-infra-system = { path = "../host-infra-system" }
