+++ apps/desktop/src-tauri/crates/host-infra-beads/src/lifecycle/mod.rs
@@ -1,6 +1,8 @@
 mod context;
 mod coordinator;
+mod error;
 mod provisioner;
 mod verifier;
 
 pub(crate) use coordinator::BeadsLifecycle;
+pub(crate) use error::LifecycleError;
